requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.24.0
matplotlib>=3.6.0
ortools>=9.7.0
openpyxl>=3.1.0
pillow>=9.5.0